fix(options): fix build error and optwin test for 'wrapcolumn'
Problem:
The 'wrapcolumn' option caused two CI failures:
1. Clang build error: implicit conversion from OptInt (long long) to int
   in plines.c win_wrap_width() triggered -Wshorten-64-to-32 with -Werror.
2. optwin_spec.lua test failure: 'wrapcolumn' was registered in
   options.lua but missing from runtime/scripts/optwin.lua, so the
   ":options shows all options" test failed.

Solution:
1. Add explicit (int) cast on w_p_wcl in win_wrap_width().
2. Add 'wrapcolumn' entry to optwin.lua next to 'wrap'.

+++ src/nvim/plines.c
@@ -415,7 +415,7 @@ int win_wrap_width(win_T *wp)
 {
   int view = wp->w_view_width;
   if (wp->w_p_wrap && wp->w_p_wcl > 0) {
-    int effective = wp->w_p_wcl + win_col_off(wp);
+    int effective = (int)wp->w_p_wcl + win_col_off(wp);
     return MIN(effective, view);
   }
   return view;
